AzureAdUser#authorities shouldn't be transient
Implementations of UserDetails are expected to be serialized in http
session.

In OSS Jenkins this may not be a problem because there is only
one node and the session is never deserialized, however in CloudBees CI HA, the session is clustered serialization/deserialization must work.

The current implementation clears any authorities upon serialization.

I'm not clear why the implementation was this way from the plugin inception.